Fix mobile application
